Fixing bug with imdb.com twitter's jot.html.
We have a form tag that creates a post request. Need to log that correctly
by adding the form element into the FormSubmission, and using that
to set the request wprof element mapping correctly.

Another bug: don't set the parent element of a resource to the computation if it's an
iframe. The iframe download happens after we hit the tag, so the computation is set incorrectly.